+++ config/me_cms.php
@@ -184,8 +184,5 @@ return ['MeCms' => [
         'reset_password' => true,
         //Signup (enabled or disabled)
         'signup' => true,
-        //Displays the userbar. This will have effect only on the frontend; in
-        //  the backend, the userbar will always be visible
-        'userbar' => true,
     ],
 ]];