+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 2.x branch
 ## 2.31 branch
 ### 2.31.6
-* removed the default `userbar` and all code related.
+* removed the default `userbar` and all code related;
+* little fixes.
 
 ### 2.31.5
 * `login` layout changed to `single-column`. This name clarifies that it can be used for various purposes (login, sign