+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # 2.x branch
 ## 2.31 branch
+### 2.31.6
+* fixed the `connector.minimal.php` file for ElFinder and the `FixElFinderCommand`: the Composer autoload is now
+  correctly loaded via absolute path;
+* removed the default `userbar` and all code related;
+* little fixes.
+
 ### 2.31.5
 * `login` layout changed to `single-column`. This name clarifies that it can be used for various purposes (login, sign
   up, but also error views);